Refactoring
Channging char[] to int[] to avoid unnecessary cast.
making startOfPackageMarker a local function variable instead of Propperty, beacause it's not important outseide the Parse method.
Adding a Parse Overload to clarify that markerSize can change and is not a constant.
Adding new variable cursor and assign it to FirtstMarer at to make clear that there is one Solution whitch sets sthe variable instead of just breaking the loop.