Massive Refactoring
in my first implementations I assumed larger projects per day, which need their own solution and namespaces to keep them apart and maintainable.

Also a stronger reuse of the code was conceivable.

Now the projects and namespaces were dissolved,

the two puzzles of the day were combined to one class with two results.